Funcionando corretamente

+++ src/sorts/HeapSort.java
@@ -18,8 +18,8 @@ public class HeapSort {
 		return result;
 	}
 
-	public long getTempo() {
-		return tempo;
+	public int getTempo() {
+		return (int)tempo;
 	}
 
 	public void sort(int arr[])
